requirements/apps.txt
-r libs.txt
factory_boy==2.1.1
requirements/src/djangowind-0.10.3.tar.gz
requirements/src/sorl-3.1.tar.gz
requirements/src/tagging-0.3-pre.tar.gz
requirements/src/typogrify-1.0.tar.gz
requirements/src/django-staticmedia-0.2.tar.gz
requirements/src/django-indexer-0.2.1.tar.gz
requirements/src/django-templatetag-sugar-0.1.tar.gz
requirements/src/django-paging-0.2.2.tar.gz
requirements/src/django-munin-0.1.1.tar.gz
requirements/src/South-0.7.5.tar.gz
requirements/src/django-annoying-0.7.6.tar.gz
requirements/src/django-nose-0.1.3.tar.gz
requirements/src/django-appconf-0.4.1.tar.gz
requirements/src/django_compressor-1.1.1.tar.gz
requirements/src/django-statsd-mozilla-0.3.3.tar.gz
requirements/src/raven-0.7.0.tar.gz
requirements/src/django-bootstrap-form-2.0.3.tar.gz
requirements/src/django-debug-toolbar-0.9.4.tar.gz
requirements/src/django-waffle-0.9.0.tar.gz
requirements/src/django-jenkins-0.14.0.tar.gz
requirements/src/django-smoketest-0.1.3.tar.gz
requirements/src/django-extensions-1.1.0.tar.gz
requirements/src/django-stagingcontext-0.1.0.tar.gz
requirements/src/django-impersonate-0.6.0.tar.gz
